requirements.txt
streamlit
opencv-python
torch
pandas
numpy
Pillow
PyYAML
ultralytics
python-docx
openpyxl
reportlab
pypandoc
odfpy
